+++ index.js
@@ -1,5 +1,0 @@
-import VGNav from "./assets/js/vgnav";
-
-export default {
-	VGNav
-}