+++ src/VGNav/js/VGNav.js
@@ -1,43 +1,54 @@
+/**
+ * --------------------------------------------------------------------------
+ * Module: VGNav
+ * Автор: Vegas DEV
+ * Лицензия: смотри LICENSE.md
+ * --------------------------------------------------------------------------
+ */
+
+/**
+ * Сайдбар нам пригодится
+ */
+import VGSidebar from "../../VGSidebar/js/VGSidebar";
+
 class VGNav {
 	constructor (arg, callback) {
 		this.settings = Object.assign({
-			expand: 'lg',
-			layout: 'sidebar',
-			isHover: false,
-			toggle: '<span class="default"></span>',
-			mobileTitle: '',
+			expand: 'lg', // Медиа точка, принцип позаимствован у https://getbootstrap.com/
+			placement: 'horizontal', // Расположение основной навигации. Либо она горизонтальная (horizontal), либо вертикальная (vertical)
+			isHover: false, // Выпадающее меню будет открываться при наведении если определено как true, или при клике если false
+			toggle: '<span class="default"></span>', // Кастомный переключатель для выпадающего списка
+			mobileTitle: '', // Помимо иконки (с полосками), можно добавить заголовок, например: "Меню" или "Навигация"
 			sidebar: {
 				placement: 'right',
-				clone: null
+				clone: null,
+				hash: false
 			}
 		}, arg);
 
+		// Функция обратного вызова
 		this.callback = callback;
+
+		// Карта медиа точек см. https://getbootstrap.com/docs/5.2/layout/grid/#grid-options
 		this.breakpoints = {
-			max: {
-				xl: 1921,
-				lg: 1200,
-				md: 992,
-				sm: 768,
-				xs: 480
-			},
-			min: {
-				xl: 1200,
-				lg: 992,
-				md: 768,
-				sm: 480,
-				xs: 0
-			}
+			xs: 0,
+			sm: 576,
+			md: 768,
+			lg: 992,
+			xl: 1200,
+			xxl: 1400
 		};
+
+		// Программные переменные
 		this.container = '.vg-nav';
+		this.sidebar = 'vg-nav-sidebar';
 		this.classes = {
-			container: 'vg-nav-main-container',
+			container: 'vg-nav-wrapper',
+			sidebar: 'vg-sidebar',
 			hamburger: 'vg-nav-hamburger',
-			sidebar: 'vg-nav-sidebar',
-			collapse: 'vg-nav-collapse',
-			overlay: 'vg-nav-overlay',
 			cloned: 'vg-nav-cloned',
 			hover: 'vg-nav-hover',
+			XXL: 'vg-nav-xxl',
 			XL: 'vg-nav-xl',
 			LG: 'vg-nav-lg',
 			MD: 'vg-nav-md',
@@ -48,11 +59,16 @@ class VGNav {
 		this.isInit = false;
 
 		if (!this.isInit) {
-			this.init();
+			this.init(this.callback);
 		}
 	}
 
-	init() {
+	/**
+	 * Вот, вот начнем
+	 * @param callback
+	 * @returns {boolean}
+	 */
+	init(callback) {
 		let _this = this,
 			$container = document.querySelector(_this.container),
 			$navigation = document.querySelector('.' + _this.classes.container);
@@ -61,11 +77,11 @@ class VGNav {
 			return false;
 		}
 
-		// Определим в основной контайнер конфигурационные классы
+		// Определим в основной контейнер конфигурационные классы
 		$container.classList.add('vg-nav-' + _this.settings.expand);
 
 		// Метод открытия меню при клике или наведении
-		if(_this.settings.isHover) {
+		if (_this.settings.isHover) {
 			$container.classList.add(_this.classes.hover);
 		}
 
@@ -78,11 +94,10 @@ class VGNav {
 		});
 
 		// Устанавливаем гамбургер
-		let responsive_class = $container.classList.contains(_this.classes.XL) || $container.classList.contains(_this.classes.LG) || $container.classList.contains(_this.classes.MD) || $container.classList.contains(_this.classes.SM) || $container.classList.contains(_this.classes.XS)
+		let responsive_class = $container.classList.contains(_this.classes.XXL) || $container.classList.contains(_this.classes.XL) || $container.classList.contains(_this.classes.LG) || $container.classList.contains(_this.classes.MD) || $container.classList.contains(_this.classes.SM) || $container.classList.contains(_this.classes.XS)
 
-		if(responsive_class) {
+		if (responsive_class) {
 			let mTitle = '';
-
 			if (_this.settings.mobileTitle) {
 				mTitle = '<span class="' + _this.classes.hamburger + '--title">'+ _this.settings.mobileTitle +'</span>';
 			}
@@ -90,49 +105,38 @@ class VGNav {
 			$container.insertAdjacentHTML('afterbegin','<a href="#" class="' + _this.classes.hamburger + '">' + mTitle + '<span class="' + _this.classes.hamburger + '--lines"><span></span><span></span><span></span></span></a>');
 		}
 
-		// Слои мобильной навигации
-		if (this.settings.layout === 'sidebar') {
-			let $sidebar = document.getElementsByClassName(_this.classes.sidebar),
-				opt_sidebar = _this.settings.sidebar || false,
-				sidebarOpen = opt_sidebar.placement || 'right',
-				$_sidebar;
+		// инит сайдбара
+		let $sidebar = document.getElementById(_this.sidebar),
+			opt_sidebar = _this.settings.sidebar || false,
+			sidebarOpen = opt_sidebar.placement || 'right',
+			historyState = opt_sidebar.hash ? 'data-hash="true"' : '';
 
-			let $collapse = document.getElementsByClassName(_this.classes.collapse);
-			if ($collapse.length) {
-				$collapse[0].remove();
-			}
+		if (responsive_class) {
+			if (!$sidebar) {
+				let mTitle = '';
+				if (_this.settings.mobileTitle) {
+					mTitle = '<span class="' + _this.classes.sidebar + '-title">'+ _this.settings.mobileTitle +'</span>';
+				}
 
-			if (responsive_class) {
-				if (!$sidebar.length) {
-					document.body.insertAdjacentHTML('beforeend','<div class="' + _this.classes.sidebar + ' ' + sidebarOpen + '">' +
-						'<div class="' + _this.classes.sidebar + '__close" data-dismiss="' + _this.classes.sidebar +'">&times;</div>' +
-						'<div class="' + _this.classes.sidebar + '__content"></div>' +
-						'</div>');
+				document.body.insertAdjacentHTML('beforeend','<div class="' + _this.classes.sidebar + ' ' + sidebarOpen + '" id="'+ _this.sidebar +'" '+ historyState +'>' +
+					'<div class="vg-sidebar-content">' +
+					'<div class="vg-sidebar-header">'+ mTitle +'<div class="' + _this.classes.sidebar + '-close" data-dismiss="' + _this.classes.sidebar +'">&times;</div></div>' +
+					'<div class="vg-sidebar-body"></div>' +
+					'</div></div>');
 
-					let $clone_target = document.getElementsByClassName(_this.classes.sidebar + '__content');
-					_this.cloneNavigation($clone_target, $container.querySelector('.' + _this.classes.container));
-				} else {
-					$_sidebar = $sidebar[0].cloneNode(true);
-					document.body.appendChild($_sidebar);
-					$sidebar[1].classList.add(sidebarOpen);
-					$sidebar[0].remove();
+				let $clone_target = document.getElementsByClassName(_this.classes.sidebar + '-body');
+				_this.cloneNavigation($clone_target, $container.querySelector('.' + _this.classes.container));
+			} else {
+				if ('clone' in opt_sidebar) {
+					if (opt_sidebar.clone) {
+						let $clone_target = document.querySelector('.' + _this.classes.sidebar).querySelectorAll(opt_sidebar.clone);
 
-					if ('clone' in opt_sidebar) {
-						if(opt_sidebar.clone) {
-							let $clone_target = document.querySelector('.' + _this.classes.sidebar).querySelectorAll(opt_sidebar.clone);
-
-							if ($clone_target) {
-								_this.cloneNavigation($clone_target, $container.querySelector('.' + _this.classes.container));
-							}
+						if ($clone_target) {
+							_this.cloneNavigation($clone_target, $container.querySelector('.' + _this.classes.container));
 						}
 					}
 				}
-
-				document.body.insertAdjacentHTML('beforeend','<div class="' + _this.classes.overlay + ' ' + sidebarOpen + '"></div>');
 			}
-		} else if (_this.settings.layout === 'collapse')  {
-			let $collapse = document.getElementsByClassName(_this.classes.collapse);
-			_this.cloneNavigation($collapse, $container.querySelector('.' + _this.classes.container))
 		}
 
 		this.isInit = true;
@@ -140,7 +144,12 @@ class VGNav {
 		this.toggle(this.callback);
 	}
 
-	toggle (callback) {
+	/**
+	 * Вкл/выкл дропа.. Тугл же
+	 * @param callback
+	 * @returns {boolean}
+	 */
+	toggle(callback) {
 		if (!this.isInit) return false;
 
 		let _this = this,
@@ -148,8 +157,7 @@ class VGNav {
 			$navigation = document.querySelectorAll('.' + _this.classes.container),
 			$click_a = document.querySelectorAll('.' + _this.classes.container + ' li > a'),
 			$click_hamburger = $container.querySelector('.' + _this.classes.hamburger),
-			$click_overlay = document.querySelector('.' + _this.classes.overlay),
-			$click_dismiss = document.querySelector('[data-dismiss=vg-nav-sidebar]');
+			$click_dismiss = document.querySelector('[data-dismiss="vg-nav"]');
 
 		// Функция обратного вызова после инициализации скрипта
 		if (callback && 'afterInit' in callback) {
@@ -158,12 +166,12 @@ class VGNav {
 
 		$click_a.forEach(function (elem) {
 			elem.onclick = function (event) {
-				if (_this.clickable()) return;
+				if (_this.clickable()) return false;
 
 				let $_self = this,
 					$li = $_self.closest('li');
 
-				clickBefore(callback, _this, event)
+				clickBefore(callback, _this, event);
 
 				// Открываем обычное меню
 				if ($li.classList.contains('dropdown')) {
@@ -237,35 +245,34 @@ class VGNav {
 			}
 		});
 
-		// Открываем, закрываем боковую панель или выпадающий список
-		$click_hamburger.onclick = function () {
-			let $_self = this;
-
-			$_self.classList.toggle('show');
-
-			if (_this.settings.layout === 'sidebar') {
-				document.getElementsByClassName(_this.classes.sidebar)[0].classList.toggle('show');
-				document.getElementsByClassName(_this.classes.overlay)[0].classList.toggle('show');
-
-				if (!document.body.classList.contains('vg-nav-sidebar-open')) {
-					let width_scrollbar = window.innerWidth - document.documentElement.clientWidth;
-
-					document.body.classList.add('vg-nav-sidebar-open');
-					document.body.style.paddingRight = width_scrollbar + 'px';
-				} else {
-					document.body.classList.remove('vg-nav-sidebar-open');
-					document.body.style.paddingRight = 0;
-				}
-			} else if (_this.settings.layout === 'collapse') {
-				document.getElementsByClassName(_this.classes.collapse)[0].classList.toggle('show');
+		if ($click_dismiss) {
+			$click_dismiss.onclick = () => {
+				_this.dispose($navigation);
+				_this.dispose($navigation, 'dropdown-mega');
+				return false;
 			}
+		}
+
+		// если меню свернулось вызываем боковую панель
+		$click_hamburger.onclick = function () {
+			let $_self = this,
+				target = _this.sidebar,
+				options = {
+					hash: _this.settings.sidebar.hash
+				};
+
+			let $sidebar = new VGSidebar(target, options);
+			$sidebar.open({
+				beforeOpen: function () {
+					$_self.classList.add('show');
+				},
+				afterClose: function () {
+					$_self.classList.remove('show');
+				}
+			})
 
 			return false;
 		}
-
-		// Альтернативы закрытию боковой папнели
-		if ($click_overlay) $click_overlay.onclick = () => { closeSidebar(); return false; };
-		if ($click_dismiss) $click_dismiss.onclick = () => { closeSidebar(); return false; };
 
 		function clickBefore(callback, $this, event) {
 			// Функция обратного вызова клика по ссылке до начала анимации
@@ -280,37 +287,13 @@ class VGNav {
 				if (typeof callback.afterClick === 'function') callback.afterClick($this, event)
 			}
 		}
-
-		function closeSidebar () {
-			let hamburger = document.getElementsByClassName(_this.classes.hamburger);
-			if (hamburger.length) {
-				hamburger[0].classList.remove('show');
-			}
-
-			if (_this.settings.layout === 'sidebar') {
-				let sidebar = document.getElementsByClassName(_this.classes.sidebar),
-					overlay = document.getElementsByClassName(_this.classes.overlay);
-
-				if (sidebar && overlay) {
-					sidebar[0].classList.remove('show');
-					overlay[0].classList.remove('show');
-				}
-
-				if(document.body.classList.contains('vg-nav-sidebar-open')) {
-					document.body.classList.remove('vg-nav-sidebar-open');
-					document.body.style.paddingRight = 0;
-				}
-
-				_this.dispose($navigation);
-				_this.dispose($navigation, 'dropdown-mega');
-			} else if (_this.settings.layout === 'collapse') {
-				document.getElementsByClassName(_this.classes.collapse)[0].classList.remove('show');
-			}
-
-			return false;
-		}
 	}
 
+	/**
+	 * Насильно прячем все что выпало....
+	 * @param $container
+	 * @param className
+	 */
 	dispose ($container, className = 'dropdown') {
 		let elements;
 
@@ -331,12 +314,21 @@ class VGNav {
 		}
 	}
 
+	/**
+	 * Клонирование навигации
+	 * @param $target_clone
+	 * @param $navigation
+	 */
 	cloneNavigation ($target_clone, $navigation) {
 		let clone_navigation = $navigation.cloneNode(true);
 		clone_navigation.classList.add(this.classes.cloned);
 		$target_clone[0].appendChild(clone_navigation);
 	}
 
+	/**
+	 * - А можно кликнуть?
+	 * @returns {boolean}
+	 */
 	clickable () {
 		let $container = document.querySelector(this.container);
 
@@ -345,23 +337,30 @@ class VGNav {
 		} else {
 			return false;
 		}
-	};
+	}
 
+	/**
+	 * - А можно кликнуть?
+	 * - Ща проверим...
+	 * @returns {boolean}
+	 */
 	checkResponsiveClass() {
 		let $container = document.querySelector(this.container);
 
-		if ($container.classList.contains(this.classes.XL)) {
-			this.current_responsive_size = this.breakpoints.max.xl;
+		if ($container.classList.contains(this.classes.XXL)) {
+			this.current_responsive_size = this.breakpoints.xxl;
+		} else if ($container.classList.contains(this.classes.XL)) {
+			this.current_responsive_size = this.breakpoints.xl;
 		} else if ($container.classList.contains(this.classes.LG)) {
-			this.current_responsive_size = this.breakpoints.max.lg;
+			this.current_responsive_size = this.breakpoints.lg;
 		} else if ($container.classList.contains(this.classes.MD)) {
-			this.current_responsive_size = this.breakpoints.max.md;
+			this.current_responsive_size = this.breakpoints.md;
 		} else if ($container.classList.contains(this.classes.SM)) {
-			this.current_responsive_size = this.breakpoints.max.sm;
+			this.current_responsive_size = this.breakpoints.sm;
 		} else if ($container.classList.contains(this.classes.XS)) {
-			this.current_responsive_size = this.breakpoints.max.xs;
+			this.current_responsive_size = this.breakpoints.xs;
 		} else {
-			this.current_responsive_size = this.breakpoints.max.xs;
+			this.current_responsive_size = this.breakpoints.xs;
 		}
 
 		return window.innerWidth >= this.current_responsive_size;